Refactor namespaces and simplify method calls
